requirements.txt
streamlit
pandas
groq
pytesseract
pillow
speechrecognition
psycopg2-binary
reportlab
matplotlib
dotenv
sqlalchemy